+++ core/build.rs
@@ -1,17 +1,6 @@
 #![allow(unused_imports)]
 use std::{env, fs, path::PathBuf};
 
-struct Cleanup {
-    to_delete: PathBuf,
-}
-
-impl Drop for Cleanup {
-    fn drop(&mut self) {
-        fs::remove_file(&self.to_delete)
-            .unwrap_or_else(|e| log::warn!("failed to remove file {:?}: {}", self.to_delete, e));
-    }
-}
-
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     env_logger::Builder::new()
         .filter(None, log::LevelFilter::Debug)
@@ -20,8 +9,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let curr_dir = env::current_dir().unwrap_or_else(|e| panic!("current directory is inaccessible: {}", e));
 
     let scalapb_target = curr_dir.join("../scalapb/scalapb.proto");
-    fs::copy(scalapb_target.parent().unwrap().join("_scalapb.proto"), &scalapb_target)?;
-    let _cleanup = Cleanup { to_delete: scalapb_target };
+    if !scalapb_target.exists() {
+        fs::copy(scalapb_target.parent().unwrap().join("_scalapb.proto"), &scalapb_target)?;
+    }
 
     let proto_files = vec!["core/protobuf/entities.proto"];
     let proto_includes = vec![".."];
